eliminate change handlers
